test all build variants
