+++ .github/workflows/deploy_dev.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Grant execute permission for gradlew
         run: chmod +x gradlew
       - name: Test with Gradle
-        run: ./gradlew test
+        run: ./gradlew check
       - name: Upload test artifacts
         if: always()
         uses: actions/upload-artifact@v4
