--- tests/src/commonTest/kotlin/com/github/klee0kai/stone/test/bindinstance/singlemethod_inject/MoonSkyProtectInjectedTests.kt
@@ -1,34 +1,0 @@
-package com.github.klee0kai.stone.test.bindinstance.singlemethod_inject
-
-import com.github.klee0kai.stone.weakref.WeakRef
-import com.github.klee0kai.test.di.bindinstance.singlemethod_inject.StarSkyComponentStoneComponent
-import com.github.klee0kai.test.mowgli.MoonSky
-import com.github.klee0kai.test.mowgli.galaxy.Mercury
-import com.github.klee0kai.test.mowgli.galaxy.Sun
-import kotlin.test.Test
-import kotlin.test.assertNull
-
-class MoonSkyProtectInjectedTests {
-
-    @Test
-    fun withoutProtectInjectedTest() {
-        //Given
-        val component = StarSkyComponentStoneComponent()
-        val mercury = WeakRef(Mercury())
-        val star = WeakRef(Sun())
-        component.starModule().star(star.get())
-        component.mercury(mercury.get())
-
-        //When
-        var moonSky: MoonSky? = MoonSky()
-        component.inject(moonSky)
-        moonSky = null
-        component.gcAll()
-
-
-        //Then
-        assertNull(mercury.get())
-        assertNull(star.get())
-    }
-
-}
