--- tests/src/commonTest/kotlin/com/github/klee0kai/stone/test/inject/HorseProtectInjectTests.kt
@@ -1,34 +1,0 @@
-package com.github.klee0kai.stone.test.inject
-
-import com.github.klee0kai.stone.weakref.Memory
-import com.github.klee0kai.stone.weakref.WeakRef
-import com.github.klee0kai.test.di.base_forest.ForestComponentStoneComponent
-import com.github.klee0kai.test.mowgli.animal.Horse
-import kotlin.test.Test
-import kotlin.test.assertNotNull
-import kotlin.test.assertNull
-
-class HorseProtectInjectTests {
-
-    @Test
-    fun withoutProtectInjectTest() {
-        // Given
-        val DI = ForestComponentStoneComponent()
-        var horse: Horse? = Horse()
-
-
-        //When
-        DI.inject(
-            horse,
-            stoneLifeCycleOwner = { }
-        )
-
-        val historyWeakReference = WeakRef(horse?.history)
-        horse = null
-        Memory.gc()
-
-        //Then: without protect all not uses should be garbage collected
-        assertNull(historyWeakReference.get())
-    }
-
-}
