+++ tests/src/jvmTest/kotlin/com/github/klee0kai/stone/test/bindinstance/singlemethod_gc/PlanetRollingTests.kt
