+++ tests/src/jvmTest/kotlin/com/github/klee0kai/stone/test/inject/SchoolProtectInjectWrappersTests.kt
