+++ tests/src/mingwX64Main/kotlin/com/github/klee0kai/test/di/bindinstance/simple/GodWorkspaceComponent.mingwX64.kt
@@ -1,0 +1,3 @@
+package com.github.klee0kai.test.di.bindinstance.simple
+
+actual fun genGodWorkspaceComponent(): GodWorkspaceComponent  = GodWorkspaceComponentStoneComponent()
