+++ weakref_multiplatform/src/wasmJsMain/kotlin/com/github/klee0kai/stone/weakref/SoftRef.js.kt
@@ -4,9 +4,9 @@ import js.WeakRef as JsWeakRef
 
 actual class SoftRef<T : Any?> actual constructor(value: T) : Ref<T?>, AutoCloseable {
 
-    var weakRef = value?.let { JsWeakRef(value) }
+    var weakRef = (value as? JsAny)?.let { JsWeakRef(value) }
 
-    actual override fun get(): T? = weakRef?.deref()
+    actual override fun get(): T? = weakRef?.deref() as T?
 
     actual fun clear() {
         weakRef = null
